UINV-22 refactoring
